src/requirements.txt
Pillow==8.2.0
Flask==2.0.1
pydocstyle==6.1.1
python-dateutil==2.8.1
python-docx==0.8.11
pytz==2021.1
pandas==1.2.5
requests==2.25.1
six==1.16.0
snowballstemmer==2.1.0
urllib3==1.26.6
Werkzeug==2.0.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
docx==0.2.4
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
numpy==1.21.0